Fix: file upload logic to middleware
